+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
     <string name="app_name">MyMedicApp</string>
-    <string name="user_name">Nombre</string>
+    <string name="user_name"> Nombre</string>
     <string name="user_email">Correo electrónico</string>
     <string name="user_password">Contraseña</string>
     <string name="confir_password">Confirmar contraseña</string>
